requirements.txt
fastapi
uvicorn
pymongo
bcrypt
python-jose[cryptography]
python-dotenv